refactor: rename agent features and chat hook for clarity
