+++ projects/u7agent/src/features/agent-service/agent-config.ts
